+++ auth/register.go
@@ -2,5 +2,6 @@ package auth
 
 import (
 	_ "github.com/godx-jp/godx-platform-framework/auth/drivers/apikey"
+	_ "github.com/godx-jp/godx-platform-framework/auth/drivers/hmac"
 	_ "github.com/godx-jp/godx-platform-framework/auth/drivers/jwt"
 )